Changed Namespaces and reorganized scripts
