+++ Assets/Scripts/SOEvents/GameEventFloat.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace GameEvents.Events {
+namespace SOEvents{
     [CreateAssetMenu(fileName = "New Game Event (float)", menuName = "Game Events/New Game Event (float)")]
     public class GameEventFloat : BaseGameEvent<float> { }
 }
